fix: tooltip cosmetic
